parse-server/requirements.txt
python-dotenv==1.0.0
TgCrypto==1.2.5
Telethon==1.33.1
asyncpg==0.29.0
asyncio==3.4.3
fastapi==0.108.0
uvicorn==0.25.0
